Move UI to internal/module
